backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-multipart
langchain
langchain-groq
langchain-huggingface
langchain-chroma
langchain-community
sentence-transformers
pypdf
docx2txt
chromadb
PyJWT
slowapi
langchain-text-splitters